requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt
supabase
cloudinary
pandas
numpy
scikit-learn
openpyxl
python-multipart
itsdangerous